librbd: return an error when removing a non-existent image
Try treating the image as new format if it's not in the old-style
directory, which is the last step in old-style removal. Then if the
image is not found in the new-style directory, -ENOENT will be
returned, preserving the semantics that existed prior to
d86a526db34a9b989891ba9b3566f30e964e02d3.

Signed-off-by: Josh Durgin <josh.durgin@inktank.com>