PG: use stats from primary after rewinding divergent entries
If the osd recieving the info has divergent entries, it will
also have a "divergent" stat structure.

Probably fixes #2769.

In cases like #2769, this bug can result in a primary with a stat
structure which double counts an operation: once for the
divergent operation, and once for the replay.

This is another way for the bug addressed in
0cb0520970fe2657398f421f6096c3c55b8a254c to happen.

Signed-off-by: Samuel Just <sam.just@inktank.com>