osdc/ObjectCacher: do not take Object ref for bh writes
This reverts commit 1144ad960272030d3f9afd6cd47e168c50486327.

There is no reason to carry a ref for the writes; it just makes things
more confusing because the refs aren't used for lifecycle, only for
LRU pinning.  We would need to duplicate the close_object() conditional
here for this to work right.

This fixes #3431, in which a slow osd reply has an Object pinned, but when
we try to truncate it away we hit the assert in can_close().

Signed-off-by: Sage Weil <sage@inktank.com>