osd: special case CALL op to not have RD bit effects
In commit c7b9f6334ce9bff105ca61b26b4693ccd54b51d7 we treat a CALL as
different from a normal "read", but we did not adjust the behavior
determined by the RD bit in the op.  We tried to fix that in
aee77d1f078d8aaca8eb6f1c869d3a8e75a24ff6, but changing the op code breaks
compatibility, so that was reverted.

Instead, special-case CALL in the helper--the only point in the code that
actually checks for the RD bit.  (And fix one lingering user to use that
helper appropriately.)

Fixes: #3731
Signed-off-by: Sage Weil <sage@inktank.com>
Reviewed-by: Dan Mick <dan.mick@inktank.com>