mds: initialize tableservers/clients on mds creation
The handle_mds_recovery(who) path initializes the anchorclients by having
the server send a 'ready' message on recovery when the server is active
and a peer becomes active.  Similarly, recovery_done() does the same when
the server becomes active.  However, this misses the creation path.  Handle
that explicitly in boot_create.

Fixes: #4619
Signed-off-by: Sage Weil <sage@inktank.com>