mon: revert part of PaxosService::is_readable() change
In 6dc7880aefd75e25899b89f0b6a28077cd031142 is_readable() was changed to
call is_active(), but that has a check for is_bootstrapping(), so there is
a semantic change.

As a result, we may fail PaxosService::is_readable() (due to bootstrapping)
and then try to call Paxos::wait_for_readable().  That will assert that
Paxos::is_readable() is false, but it will be true and we will crash.

Revert that part of the change, since the semantic change was not
intentional.

Signed-off-by: Sage Weil <sage@inktank.com>
Reviewed-by: Greg Farnum <greg@inktank.com>