mon/PaxosService: allow paxos service writes while paxos is updating
In commit d34f80bf1853de04fc4551770dbe23025f68c34e I mistakenly made
is_writeable() false while paxos was updating due to a misread of
Paxos::propose_new_value() (I didn't see that it would queue).
This is problematic because it narrows the window during which each service
is writeable for no reason.

Allow service to be writeable both when paxos is active and updating.

Signed-off-by: Sage Weil <sage@inktank.com>